requirements.txt
opencv-python==4.5.3.56
ttkthemes==3.2.2
pipwin==0.5.1
bcrypt==3.2.0
mysql-connector-python==8.0.26
pyinstaller==4.7
pygrabber==0.1
comtypes==1.1.10
numpy
pandas
scikit-learn
mediapipe